Add files via upload
This program only makes a rough estimate of all calculations and uses only /4 for leap years calculations as i did not wanted to heavily rely on datetime module. Even though the calculation will be off but they will be off by a few days only. It also calculates weeks, hours, minutes and seconds as well.